Updates...more

+++ ocr/deployFunction.ps1
@@ -1,6 +1,6 @@
 [cmdletbinding()]
 param(
-    
+    [switch]$preservePreviousResourceGroups
 )
 
 $location = 'West US 2'
@@ -17,72 +17,60 @@ $ErrorActionPreference = "Stop"
 
 if ((get-command -module "az*").count -eq 0) {
     write-verbose "Couldn't find any Az modules, so I'm installing them"
-    write-progress -Activity "Installing Azure management modules"
     Install-Module -Name Az -AllowClobber -Scope CurrentUser -Force
     Import-Module Az
 }
 
 try {
-    write-progress "Testing user authentication with Azure"
+    write-verbose "Testing user authentication with Azure"
     Get-AzResourceGroup | out-null
 } catch {
     Connect-AzAccount
 }
 
 $resourceGroups = Get-AzResourceGroup
-Write-Verbose "Removing any previous resource groups from this project..."
-$resourceGroups | Where-Object {$_.ResourceGroupName -like "$resourceGroupRoot*"} | Remove-AzResourceGroup -Verbose -AsJob -Force
 
-$resourceGroup = $resourceGroups | Where-Object { $_.ResourceGroupName -eq $resourceGroupName }
-if ( $null -eq $resourceGroup) {
-    write-verbose "Creating resource group"
-    Write-Progress -Activity "Creating resource group $resourceGroupName"
-    New-AzResourceGroup -Name $resourceGroupName -Location $location -force
+if($preservePreviousResourceGroups) {
+    Write-Warning "Preserving previous resource groups...they're still out there..."
+} else {
+    Write-Verbose "Removing any previous resource groups from this project..."
+    $resourceGroups | Where-Object {$_.ResourceGroupName -like "$resourceGroupRoot*"} | Remove-AzResourceGroup -Verbose -AsJob -Force
 }
 
-# $storageAccounts = Get-AzStorageAccount
-# $storageAccount = $storageAccounts | Where-Object {$_.StorageAccountName -eq $storageAccountName}
-# if ($null -eq $storageAccount)
-# {
-#     write-host "Creating storage account"
-#     Write-Progress -Activity "Creating storage account $storageAccountName"
-#     New-AzStorageAccount -ResourceGroupName $resourceGroupName -AccountName $storageAccountName -Location $location -SkuName “Standard_LRS”
-# }
+Write-Verbose "Creating resource group $resourceGroupName"
+New-AzResourceGroup -Name $resourceGroupName -Location $location -force
 
-write-host "Looking for web app (function)"
-$functionAppResource = Get-AzResource | Where-Object { $_.ResourceName -eq $functionAppName -And $_.ResourceType -eq ‘Microsoft.Web/Sites’ }
+# Write-Verbose "Creating storage account $storageAccountName"
+# New-AzStorageAccount -ResourceGroupName $resourceGroupName -AccountName $storageAccountName -Location $location -SkuName “Standard_LRS”
 
-if ($null -eq $functionAppResource) {
-    # https://clouddeveloper.space/2017/10/26/deploy-azure-function-using-powershell/
+write-verbose "Creating AzFunctions web app $functionAppName"
 
-    write-verbose "Creating web app (function)"
-
-    # Create the parameters for the file, which for this template is the function app name.
-    $TemplateParams = @{
-        "subscriptionId" = (Get-AzSubscription).Id;
-        "name" = $functionAppName;
-        "location" = "West US 2";
-        "hostingEnvironment" = "";
-        "hostingPlanName" = $resourceGroupRoot+"HostingPlan"+$salt;
-        "serverFarmResourceGroup" = "serverFarmResourceGroup"+$salt;
-        "alwaysOn" = $false;
-        "storageAccountName" = $storageAccountName;
-        "linuxFxVersion" = "DOCKER|microsoft/azure-functions-dotnet-core2.0:2.0";
-        "sku" = "Dynamic";
-        "skuCode" = "Y1";
-        "workerSize" = "0";
-        "workerSizeId" = "0";
-        "numberOfWorkers"="1";
-        #"nameFromTemplate" = $functionAppName;
-    }
-
-    # Deploy the template
-    $valid = test-AzResourceGroupDeployment -ResourceGroupName $resourceGroupName -TemplateFile template.json -TemplateParameterObject $TemplateParams -Verbose
-    if($valid) {
-        New-AzResourceGroupDeployment -Name "functionDeployment" -ResourceGroupName $resourceGroupName -TemplateFile template.json -TemplateParameterObject $TemplateParams -Verbose
-    }
-    write-host "Function app created!"
+$templateParameters = @{
+    "subscriptionId" = (Get-AzSubscription).Id;
+    "name" = $functionAppName;
+    "location" = "West US 2";
+    "hostingEnvironment" = "";
+    "hostingPlanName" = $resourceGroupRoot+"HostingPlan"+$salt;
+    "serverFarmResourceGroup" = "serverFarmResourceGroup"+$salt;
+    "alwaysOn" = $false;
+    "storageAccountName" = $storageAccountName;
+    "linuxFxVersion" = "DOCKER|microsoft/azure-functions-dotnet-core2.0:2.0";
+    "sku" = "Dynamic";
+    "skuCode" = "Y1";
+    "workerSize" = "0";
+    "workerSizeId" = "0";
+    "numberOfWorkers"="1";
 }
+
+Write-Verbose "Testing az resource group deployment"
+$valid = test-AzResourceGroupDeployment -ResourceGroupName $resourceGroupName -TemplateFile template.json -TemplateParameterObject $templateParameters -Verbose
+if($valid) {
+    Write-Verbose "Deploying azure functions web app"
+    New-AzResourceGroupDeployment -Name "deploy1" -Mode Complete -ResourceGroupName $resourceGroupName -TemplateFile template.json -TemplateParameterObject $templateParameters -Verbose -DeploymentDebugLogLevel All
+}
+write-host "Function app created!"
+
+return;
 
 write-verbose "Getting storage account key"
 $keys = Get-AzStorageAccountKey -ResourceGroupName $resourceGroupName -AccountName $storageAccount
